# clean-folder/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VayuIQ Air Quality Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AqiData {
  aqi: number;
  category: string;
  dominantPollutant: string;
  temperature: number;
  humidity: number;
  windSpeed: number;
  location: string;
  lat: number;
  lng: number;
  riskPercentage: number;
  healthAdvice: string;
  timestamp: string;
}

export interface DayPrediction {
  date: string;
  aqi: number;
  category: string;
}

export interface AqiPrediction {
  predictions: DayPrediction[];
  insight: string;
}

export interface Hotspot {
  id: number;
  lat: number;
  lng: number;
  aqi: number;
  name: string;
  severity: string;
}

export interface Report {
  id: number;
  location: string;
  description: string;
  severity: string;
  status: string;
  createdAt: string;
  lat?: number;
  lng?: number;
}

export interface CreateReport {
  location: string;
  description: string;
  severity: string;
  lat?: number;
  lng?: number;
}

export interface Task {
  id: number;
  title: string;
  description?: string;
  status: string;
  notes?: string;
  location?: string;
  createdAt: string;
}

export interface CreateTask {
  title: string;
  description?: string;
  location?: string;
}

export interface UpdateTask {
  status?: string;
  notes?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
  role?: string;
}

export interface User {
  id: number;
  name: string;
  email?: string;
  role: string;
}

export interface LoginResponse {
  success: boolean;
  user?: User;
  role: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetCurrentAqiParams = {
  lat?: number;
  lng?: number;
};