# clean-folder/lib/api-zod/src/generated/types/createTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VayuIQ Air Quality Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTask {
  title: string;
  description?: string;
  location?: string;
}