# clean-folder/lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VayuIQ Air Quality Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  title: string;
  description?: string;
  status: string;
  notes?: string;
  location?: string;
  createdAt: string;
}